Try bundler.
I couldn't do the right thing and .gitignore vendor/gems (while keeping
vendor/gems/cache) because the shared host I'm deploying to is stuck back on
rubygems 1.3.1, which is pre-gem plugins (necessary for bundler).

It's also a bummer that the latest chronic was built before the concept of
separate development and runtime dependencies for gems, so hoe, rubyforge,
json_pure and rake all get pulled in needlessly.

On balance, though, it's fun for me to play with bundler. :-)