Temporarily disable enclosure downloading.
I'm getting a lot of noise because the filter runs for EVERY item in any
updated feed. So I queue up old things to download that I've already
gotten.

Maybe I can fix this with a time threshold, or maybe keeping track on my
own of what entries I've already seen. Or maybe using a planet filter
isn't the right way to do this...